feat: use appData for Cv section and modify the input layout and logic
